+++ changelog.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added checkdata.py helper script
 - Added serial number for internet gateway
 - Added label message type
+- Added IP address and wifi signal level for internet gateway
+- Added comms status for internet gateway
 
 ### Removed
 ### Changed